+++ packages/awgmi/vitest.config.ts
@@ -1,6 +1,14 @@
 import { defineConfig } from 'vitest/config'
+import { resolve } from 'path'
+
+const r = (p: string) => resolve(__dirname, p)
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      '@pancakeswap/awgmi/core': r('./core/src/index.ts'),
+    },
+  },
   test: {
     coverage: {
       reporter: ['text', 'json', 'html'],