+++ packages/token-lists/react/index.test.ts
@@ -1,0 +1,21 @@
+import { expect, test } from 'vitest'
+import * as exports from './index'
+
+test('exports', () => {
+  expect(Object.keys(exports)).toMatchInlineSnapshot(`
+    [
+      "NEW_LIST_STATE",
+      "createTokenListReducer",
+      "fetchTokenList",
+      "addList",
+      "removeList",
+      "enableList",
+      "disableList",
+      "acceptListUpdate",
+      "rejectVersionUpdate",
+      "updateListVersion",
+      "createListsAtom",
+      "useFetchListCallback",
+    ]
+  `)
+})